# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: فزعة FAZAAH – منصة الخدمات والمهنيين في اليمن
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /providers/{id}/reviews:
    get:
      operationId: getProviderReviews
      tags: [reviews]
      summary: Get reviews for a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
  /subscriptions/checkout:
    post:
      operationId: createSubscriptionPayment
      tags: [subscriptions]
      summary: Submit a manual wallet payment for a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionPaymentInput"
      responses:
        "201":
          description: Payment submitted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPayment"

  # ─── ADVERTISEMENTS ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SubscriptionPayment:
      type: object
      required: [id, providerId, plan, wallet, transactionReference, status, createdAt]
      properties:
        id:
          type: integer
        providerId:
          type: integer
        plan:
          type: string
          enum: [monthly, yearly]
        wallet:
          type: string
          enum: [jeeb, floosk, jawali, cash, one_cash, hasib, easy]
        transactionReference:
          type: string
        receiptUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, expired, refunded]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
    SubscriptionPaymentInput:
      type: object
      required: [plan, wallet, transactionReference]
      properties:
        plan:
          type: string
          enum: [monthly, yearly]
        wallet:
          type: string
          enum: [jeeb, floosk, jawali, cash, one_cash, hasib, easy]
        transactionReference:
          type: string
          minLength: 3
        receiptUrl:
          type: ["string", "null"]
...
    Review:
      type: object
      required: [id, clientId, providerId, rating, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: string
        clientAvatarUrl:
          type: ["string", "null"]
        providerId:
          type: integer
        requestId:
          type: ["integer", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [providerId, requestId, rating]
      properties:
        providerId:
          type: integer
        requestId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]

    # ── FAVORITES ──────────────────────────────────────────────────────────────
...