# lib/api-zod/src/generated/types/subscriptionPaymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * فزعة FAZAAH – منصة الخدمات والمهنيين في اليمن
 * OpenAPI spec version: 0.1.0
 */
import type { SubscriptionPaymentInputPlan } from './subscriptionPaymentInputPlan';
import type { SubscriptionPaymentInputWallet } from './subscriptionPaymentInputWallet';

export interface SubscriptionPaymentInput {
  plan: SubscriptionPaymentInputPlan;
  wallet: SubscriptionPaymentInputWallet;
  /** @minLength 3 */
  transactionReference: string;
  /** @nullable */
  receiptUrl?: string | null;
}